fix: 🐛 Fixed event definition in the MockWrappedERC20

+++ contracts/test/MockWrappedERC20.sol
@@ -9,8 +9,8 @@ import '@openzeppelin/contracts/token/ERC20/extensions/draft-ERC20Permit.sol';
 import '@openzeppelin/contracts/token/ERC20/extensions/ERC20FlashMint.sol';
 
 contract MockWrappedERC20 is ERC20, ERC20Burnable, Pausable, AccessControl, ERC20Permit, ERC20FlashMint {
-  event  Deposit(address indexed dst, uint wad);
-  event  Withdrawal(address indexed src, uint wad);
+  event Deposit(address indexed dst, uint256 wad);
+  event Withdrawal(address indexed src, uint256 wad);
 
   constructor() ERC20('MockWERC20', 'WMTK') ERC20Permit('MockWERC20') {}
